requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-groq>=0.2.0
langchain-community>=0.3.0
# anthropic>=0.40.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
fastapi>=0.115.0
uvicorn>=0.32.0
python-multipart>=0.0.12
pydantic>=2.9.0
pandas>=2.2.0
plotly>=5.24.0
dash>=2.18.0
PyPDF2>=3.0.0
python-docx>=1.1.0
SQLAlchemy>=2.0.0
aiosqlite>=0.20.0
python-dotenv>=1.0.0
httpx>=0.27.0
tiktoken>=0.8.0
rich>=13.9.0
typer>=0.13.0
